Usuario y Rol. Login con los usuarios de la base de datos. Script SQL (import.sql) se ejecuta siempre que pones play, ahi se crea un usuario para probar.

+++ src/main/java/ar/com/utn/models/PersistentEntity.java
@@ -11,15 +11,21 @@ import javax.persistence.MappedSuperclass;
 @MappedSuperclass
 public class PersistentEntity {
     @Id
-    @GeneratedValue(strategy= GenerationType.AUTO)
-    protected Integer id;
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    protected Long id;
 
     /* Setter methods */
-    public void    setId(Integer id) { this.id = id; }
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     /* Getter methods */
-    public Integer getId()           { return id; }
+    public Long getId() {
+        return id;
+    }
 
-    public boolean isNew()           { return id == null; }
+    public boolean isNew() {
+        return id == null;
+    }
 
 }
